requirements.txt
schedule
PyTelegramBotApi
python-dotenv